+++ react/components/Popup.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
-
 import Arrow from '../images/Arrow'
 
 const { Provider, Consumer } = React.createContext()
@@ -72,16 +71,18 @@ export default class Popup extends Component {
       PropTypes.func,
     ]),
     renderFooter: PropTypes.func,
+    icon: PropTypes.node,
   }
 
   static defaultProps = {
     renderFooter: () => null,
+    icon: <Arrow size={16} />
   }
 
   contentRef = React.createRef()
 
   render() {
-    const { children, renderFooter, title, id } = this.props
+    const { children, renderFooter, title, id, icon } = this.props
 
     return (
       <Consumer>
@@ -89,7 +90,7 @@ export default class Popup extends Component {
           const { isOpen, onToggle } = contextProps
           const open = isOpen(id)
 
-          const className = classNames('vtex-filter-popup relative', {
+          const className = classNames('vtex-filter-popup relative  justify-center flex ', {
             'vtex-filter-popup--open': open,
           })
 
@@ -113,16 +114,17 @@ export default class Popup extends Component {
 
           return (
             <div className={className} ref={this.contentRef}>
+            
               <button
-                className={classNames('vtex-filter-popup__button pa5 mv0 mh5 pointer flex justify-center items-center', {
+                className={classNames('vtex-filter-popup__button mv0 pointer flex justify-center items-center', {
                   'bb b--muted-1': open,
                   'bn': !open,
                 })}
                 onClick={onToggle(id)}
               >
-                <span className="vtex-filter-popup__title t-action-small ml-auto">{title}</span>
-                <span className="vtex-filter-popup__arrow-icon ml-auto">
-                  <Arrow size={8} />
+                <span className="vtex-filter-popup__title c-on-base t-action--small ml-auto">{title}</span>
+                <span className="vtex-filter-popup__arrow-icon ml-auto pl3 pt2">
+                  {icon}
                 </span>
               </button>
               <div className={contentClassName} style={{ top: contentTop }}>