+++ react/images/FilterIcon.js
@@ -2,34 +2,34 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 /**
- * Arrow icon component in svg
+ * Filter icon component in svg
  */
-const Arrow = ({ size, fillColor }) => {
+const FilterIcon = ({ size, fillColor }) => {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
       xmlnsXlink="http://www.w3.org/1999/xlink"
       width={size}
       height={size}
-      viewBox="0 0 11 6"
+      viewBox="0 0 17 17"
       fill="none"
       color={fillColor}
     >
-      <use href="#arrow-up-search-result" xlinkHref="arrow-up-search-result" />
+      <use href="#search-result-filter" xlinkHref="#search-result-filter" />
     </svg>
   )
 }
 
-Arrow.propTypes = {
+FilterIcon.propTypes = {
   /* Percentage size of the icon */
   size: PropTypes.number,
   /* Fill color for the icon */
   fillColor: PropTypes.string,
 }
 
-Arrow.defaultProps = {
+FilterIcon.defaultProps = {
   size: 20,
-  fillColor: '#3F3F40',
+  fillColor: 'currentColor',
 }
 
-export default Arrow
+export default FilterIcon